+++ TexasHoldEm/src/misc/misc.h
@@ -1,11 +1,12 @@
 #pragma once
 
 #include <stdlib.h>
+#include <stdint.h>
 #include <string>
 #include <iostream>
 #include <chrono>
 #include <thread>
-#include "../game/draw/draw.h"
+#include "draw.h"
 
 using namespace std;
 
@@ -13,7 +14,7 @@ namespace Misc {
 	void handleExit();
 	void handleEnteringGame();
 	void timer(
-		int seconds,
+		uint8_t seconds,
 		string description = "Timer will end"
 	);
 }
